+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## v0.1.0
 - Listing Blobs of some folder and container 
 - Size calculation of individual blobs and folders
+- Two implementations: AzureBlobService and AzureDatalakeService